requirements.txt
cryptography==3.4.8
bcrypt==3.1.7
python-dotenv==0.19.0